feat: Update OrderEntity and OrderItemEntity classes to use Long for customerId
The code changes modify the OrderEntity and OrderItemEntity classes to use Long instead of String for the customerId field. This ensures consistency and improves data integrity.

+++ src/main/java/br/com/rafael/orderms/listener/dto/OrderItemEvent.java
@@ -1,8 +1,10 @@
 package br.com.rafael.orderms.listener.dto;
 
+import java.math.BigDecimal;
+
 public record OrderItemEvent(
         String produto,
         Integer quantidade,
-        Double preco
+        BigDecimal preco
 ) {
 }
